Resolved bugs with non-orthogonal periodic matching. Resolved issues with setting Dirichlet boundary conditions on vself bins. Tested and matched with previous results. Tested on hcp Mg primitive cell non-orthogonal- got 1e-6 Hartree match with ABINIT energy, max force  1e-7 Hartee/Bohr ABINIT max force 0. Default linking is still to older version dealii though which has a bug for hangin node periodic matching. Switch to gcc CMakeLists to use dealii develop version

+++ include/dft.h
@@ -485,7 +485,7 @@ class dftClass
   std::map<int,std::set<int> > d_bins;
   std::vector<std::vector<int> > d_imageIdsInBins;
   std::vector<std::map<dealii::types::global_dof_index, int> > d_boundaryFlag;
-  std::vector<std::map<dealii::types::global_dof_index, int> > d_vselfBinField;
+  std::vector<std::map<dealii::types::global_dof_index, double> > d_vselfBinField;
   std::vector<std::map<dealii::types::global_dof_index, int> > d_closestAtomBin;
   std::vector<vectorType> d_vselfFieldBins;//required for configurational force
   //
